# Speed up music index generation