--- a/jpsonic-main/src/main/java/com/tesshu/jpsonic/dao/MediaFileDao.java
+++ b/jpsonic-main/src/main/java/com/tesshu/jpsonic/dao/MediaFileDao.java
@@ -982,6 +982,20 @@ public class MediaFileDao {
                 """, rowMapper, args);
     }
 
+    public List<MediaFile> getSingleSongs(List<MusicFolder> folders) {
+        if (folders.isEmpty()) {
+            return Collections.emptyList();
+        }
+        Map<String, Object> args = Map.of("types", List.of(MediaType.DIRECTORY.name(), MediaType.ALBUM.name()),
+                "folders", MusicFolder.toPathList(folders));
+        return template.namedQuery("select " + prefix(QUERY_COLUMNS, "m_file") + """
+                from media_file m_file
+                join music_folder m_folder on m_file.folder = m_folder.path
+                where type not in (:types) and parent_path in(:folders)
+                order by m_folder.folder_order, m_file.media_file_order
+                """, rowMapper, args);
+    }
+
     static class RandomSongsQueryBuilder {
 
         private final RandomSearchCriteria criteria;