--- a/jpsonic-main/src/main/java/com/tesshu/jpsonic/domain/Artist.java
+++ b/jpsonic-main/src/main/java/com/tesshu/jpsonic/domain/Artist.java
@@ -143,6 +143,7 @@ public final class Artist implements Orderable, ArtistIndexable {
         this.order = order;
     }
 
+    @Override
     public String getMusicIndex() {
         return musicIndex;
     }