--- a/jpsonic-main/src/main/java/com/tesshu/jpsonic/domain/ArtistIndexable.java
+++ b/jpsonic-main/src/main/java/com/tesshu/jpsonic/domain/ArtistIndexable.java
@@ -29,4 +29,6 @@ public sealed interface ArtistIndexable permits Artist, MediaFile, ArtistSortCan
 
     @Nullable
     String getReading();
+
+    String getMusicIndex();
 }