--- a/jpsonic-main/src/main/java/com/tesshu/jpsonic/domain/MusicFolderContent.java
+++ b/jpsonic-main/src/main/java/com/tesshu/jpsonic/domain/MusicFolderContent.java
@@ -26,21 +26,19 @@ import java.util.SortedMap;
 
 public class MusicFolderContent {
 
-    private final SortedMap<MusicIndex, List<MusicIndex.SortableArtistWithMediaFiles>> indexedArtists;
+    private final SortedMap<MusicIndex, List<MediaFile>> indexedArtists;
     private final List<MediaFile> singleSongs;
 
-    public MusicFolderContent(SortedMap<MusicIndex, List<MusicIndex.SortableArtistWithMediaFiles>> indexedArtists,
-            List<MediaFile> singleSongs) {
+    public MusicFolderContent(SortedMap<MusicIndex, List<MediaFile>> indexedArtists, List<MediaFile> singleSongs) {
         this.indexedArtists = indexedArtists;
         this.singleSongs = singleSongs;
     }
 
-    public SortedMap<MusicIndex, List<MusicIndex.SortableArtistWithMediaFiles>> getIndexedArtists() {
+    public SortedMap<MusicIndex, List<MediaFile>> getIndexedArtists() {
         return indexedArtists;
     }
 
     public List<MediaFile> getSingleSongs() {
         return singleSongs;
     }
-
 }