--- a/jpsonic-main/src/main/java/com/tesshu/jpsonic/service/MusicIndexService.java
+++ b/jpsonic-main/src/main/java/com/tesshu/jpsonic/service/MusicIndexService.java
@@ -4,6 +4,7 @@ package com.tesshu.jpsonic.service;
 import java.util.List;
 import java.util.SortedMap;
 
+import com.tesshu.jpsonic.domain.Artist;
 import com.tesshu.jpsonic.domain.MediaFile;
 import com.tesshu.jpsonic.domain.MusicFolder;
 import com.tesshu.jpsonic.domain.MusicFolderContent;
@@ -24,7 +25,7 @@ public interface MusicIndexService {
     /**
      * @since Airsonic
      */
-    SortedMap<MusicIndex, List<MusicIndex.SortableArtistWithArtist>> getIndexedId3Artists(List<MusicFolder> folders);
+    SortedMap<MusicIndex, List<Artist>> getIndexedId3Artists(List<MusicFolder> folders);
 
     /**
      * @since Airsonic