--- a/jpsonic-main/src/main/webapp/WEB-INF/jsp/home.jsp
+++ b/jpsonic-main/src/main/webapp/WEB-INF/jsp/home.jsp
@@ -190,11 +190,9 @@ $(document).ready(function(){
             <ul class="anchorList">
                 <c:forEach items="${entry.value}" var="artist" varStatus="loop">
                     <sub:url value="main.view" var="mainUrl">
-                        <c:forEach items="${artist.mediaFiles}" var="mediaFile">
-                            <sub:param name="id" value="${mediaFile.id}"/>
-                        </c:forEach>
+                        <sub:param name="id" value="${artist.id}"/>
                     </sub:url>
-                    <li><a target="main" href="${mainUrl}" title="${artist.sortableName}">${fn:escapeXml(artist.name)}</a></li>
+                    <li><a target="main" href="${mainUrl}" title="${artist.artistReading}">${fn:escapeXml(artist.name)}</a></li>
                 </c:forEach>
             </ul>
         </details>