… jpsonic-main/src/main/java/com/tesshu/jpsonic/service/MusicIndexServiceUtils.java
… jpsonic-main/src/test/java/com/tesshu/jpsonic/domain/JpsonicComparatorsTest.java
… jpsonic-main/src/test/java/com/tesshu/jpsonic/domain/JpsonicComparatorsTestUtils.java
… jpsonic-main/src/test/java/com/tesshu/jpsonic/service/MusicIndexServiceUtilsTest.java
… jpsonic-main/src/test/java/com/tesshu/jpsonic/service/scanner/MediaScannerServiceImplTest.java
… jpsonic-main/src/test/java/com/tesshu/jpsonic/service/scanner/MusicIndexServiceImplTest.java