Prerequisites: #2423

### Overview

Design improvements will make the following service methods faster:

 - MusicIndexService#getMusicFolderContent
 - MusicIndexService#getIndexedId3Artists

Some internal classes defined in Subsonic/Airsonic's MusicIndex class are no longer needed. They will be removed in this Pull Request.

### Details


getMusicFolderContent and getIndexedId3Artists are methods used in APIs, web pages, and UPnP to create music-index data objects. Music-index generation is a typical example where impedance mismatch is likely to occur. Simply put, in Subsonic and Airsonic it is done in Java, but in Jpsonic it is done in SQL. The difference is whether complex data structures are handled as objects or relationally. Typically, if performance is required, an SQL implementation is almost always chosen.

If we want to do this in SQL, the Subsonic and Airsonic DB schemas are poorly designed. At least the following two fields are required.

 - order
 - music_index

These fields must have appropriate values registered. In Latin, the article is the first thing that comes to mind. Non-Latin languages may require additional sorting tags and normalization. And, In reality, data cleansing and correction of missing data will also be necessary. 

( ... If the data integrity of hundreds of thousands of files could be imposed on the user, there would be no need for data correction, but such an idea is not very modern 😏😏😏)

In Jpsonic these are resolved during the scan. 



### Effect of improvement

The target value was determined in advance.

> Keep the index generation response for a library of 400,000 songs below 400ms

- #2346

The first access is especially slow (before being optimized by JVM). After scanning 400,000 songs on DS220+, rebooting and logging in, the speed of getMusicFolderContent will be as follows.

#### v112.2.8

> 6,493 ms
> 2,126 ms
> 1,977 ms
> 1,843 ms
> 1,625 ms
> 1,522 ms
> 1,548 ms
> 1,555 ms

#### This Pull Request

893 ms
228 ms
179 ms
177 ms
164 ms
164 ms
 
Let's say it's mostly passed...
 
 
### Non-Goal
 
 UPnP needs further improvements, but is not included in this pull request.